#88 -feat: 알림창 수정
