+++ src/components/Calendar/History/HistoryComparison.tsx
@@ -1,9 +1,9 @@
-import { HistoryProps } from '@/pages/Calendar';
 import * as S from '@/styles/Calendar/History/HistoryComparison.style';
 import ComparisonPrevMonth from '@/components/Calendar/History/Comparison/ComparisonPrevMonth';
 import ComparisonToday from '@/components/Calendar/History/Comparison/ComparisonToday';
 import HistoryDate from '@/components/Calendar/History/HistoryDate';
 import { useAppSelector } from '@/hooks/store';
+import { HistoryProps } from '@/components/Layout/Route';
 
 function HistoryComparison({ history }: HistoryProps) {
   const calendar = useAppSelector((state) => state.calendar);
