+++ src/pages/Calendar.tsx
@@ -2,54 +2,11 @@ import CalendarHeader from '@/components/Calendar/Header/CalendarHeader';
 import CalendarHistory from '@/components/Calendar/History/CalendarHistory';
 import CalendarMain from '@/components/Calendar/Main/CalendarMain';
 import * as S from '@/styles/Calendar/Calendar.style';
-import { useEffect, useState } from 'react';
-import { baseAPI } from '@/api/customAxios';
 import Income from '@/components/Calendar/Amount/Income';
 import Expenditure from '@/components/Calendar/Amount/Expenditure';
-import { useAppDispatch, useAppSelector } from '@/hooks/store';
-import { onToggle } from '@/store/historySlice';
+import { HistoryProps } from '@/components/Layout/Route';
 
-export interface History {
-  id: number;
-  userId: number;
-  year: number;
-  month: number;
-  date: number;
-  value: string;
-  category: string;
-  amount: number;
-  content: string;
-}
-
-export interface HistoryProps {
-  history: History[];
-}
-
-function Calendar() {
-  const get = useAppSelector((state) => state.history);
-  const dispatch = useAppDispatch();
-  const [history, setHistory] = useState<History[]>([]);
-
-  const getHistory = async () => {
-    try {
-      const res = await baseAPI.get('/historys');
-      setHistory(
-        res.data.filter(
-          (el: History) =>
-            el.userId ===
-            (localStorage.user ? JSON.parse(localStorage.user).userId : 0)
-        )
-      );
-      dispatch(onToggle(false));
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
-  useEffect(() => {
-    get && getHistory();
-  }, [get]);
-
+function Calendar({ history }: HistoryProps) {
   return (
     <S.CalendarContainer>
       <div className="calendarTop">
